+++ src/config/enum.ts
@@ -34,6 +34,13 @@ enum JWT_TIME {
   _15_MINUTES = '15m',
 }
 
+enum EXPIRE_TIME {
+  _30_MINUTES = 1800,
+  _60_MINUTES = 3600,
+  _15_MINUTES = 900,
+  _5_MINUTES = 300,
+}
+
 export {
   TOKEN_EXCEPTION,
   JWT_TIME,
@@ -41,4 +48,5 @@ export {
   CLOUDINARY_FOLDER_NAME,
   BASE_APIS_URL,
   SWAGGER_TAGS,
+  EXPIRE_TIME,
 };