+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # [1.1.0](https://github.com/victory-aime/my-immo-backend/compare/v1.0.0...v1.1.0) (2026-02-03)
 
-
 ### Features
 
-* **email verification settings:** email verification ([3511a69](https://github.com/victory-aime/my-immo-backend/commit/3511a6985264306563bb7dc9c93a115fc8f8eda4))
+- **email verification settings:** email verification ([3511a69](https://github.com/victory-aime/my-immo-backend/commit/3511a6985264306563bb7dc9c93a115fc8f8eda4))
 
 # 1.0.0 (2026-01-27)